part-2/requirements.txt
# Snowflake connector
snowflake-connector-python==3.7.0
snowflake-sqlalchemy==1.5.3

# Data processing
pandas==2.1.4
openpyxl==3.1.2   
numpy==1.26.3

# dbt
dbt-core==1.7.4
dbt-snowflake==1.7.2

# Utilities
python-dotenv==1.0.0
faker==22.0.0          
pyarrow==14.0.2       
schedule==1.2.1
